requirements.txt
python-dotenv>=1.0.0
colorama==0.4.6
pygame-ce==2.5.7
environs==11.0.0
pytest>=9.0.2